# docs/api/server.openapi.yaml
openapi: 3.0.3
info:
  title: 通用企业知识库平台 — web ↔ server API（v0.2 草稿）
  version: 0.2.0-draft
  description: |
    机器可读权威契约。**状态：评审中，未冻结**。本文件取代 v0.1 版本并覆盖 v0.2 设计能力；
    尚未覆盖的字段以 `TBD(评审确认)` 标注，不得据此生成生产代码。
    设计依据：docs/02-概要设计.md、docs/03-详细设计.md §16、docs/modules/enterprise-generalization/requirements。
servers:
  - url: https://api.ragkb.example.com
security:
  - oidcSession: []      # OIDC 生产：BFF 设置的 HttpOnly 会话 cookie（现状不动）
  - bearerJwt: []        # form 开发：JWT access token（Authorization: Bearer <token>）
...
components:
  securitySchemes:
    oidcSession:
      type: apiKey
      in: cookie
      name: ragkb_session
      description: BFF 设置的 Secure/HttpOnly/SameSite 会话 cookie；浏览器不保存长期 token（02-概要 §4.1）。
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        form 模式的 JWT access token；前端仅内存持有（不落 localStorage），随请求头
        `Authorization: Bearer <token>` 发送。刷新凭证走 HttpOnly cookie `ragkb_refresh`
        （SameSite=Lax, Path=/api/v1/auth，生产加 Secure），不进入请求头；刷新见
        `POST /api/v1/auth/refresh`。
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: ApiKey
      description: '机器访问；格式 `Authorization: Bearer <apikey>`，scope 限制于 key 本身。'
    oidc:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://idp.example.com/authorize
          tokenUrl: https://idp.example.com/token
          scopes:
            openid: OpenID Connect
            profile: 用户基础资料
            email: 邮箱
  parameters:
    page:
      name: page
      in: query
      description: 页号（从 1 开始）
      schema: { type: integer, minimum: 1, default: 1 }
    size:
      name: size
      in: query
      description: 每页条数
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    kbPath:
      name: kbId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    idempotencyKey:
      name: Idempotency-Key
      in: header
      description: 幂等键，写操作必填；同一键 + 相同请求摘要返回原结果，不同摘要返回 409
      schema: { type: string, maxLength: 128 }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: "0" }
        message: { type: string }
        data: {}
    Error:
      type: object
      required: [code, message, requestId, timestamp]
      properties:
        code: { type: string, example: "E-1003" }
        message: { type: string }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
        details: { type: object, nullable: true }
    PageData:
      type: object
      properties:
        items: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
        hasMore: { type: boolean }
...
    Task:
      type: object
      properties:
        taskId: { type: string }
        type: { type: string, description: "如 INDEX_BUILD / CLONE / DELETION / SYNC / PARSE" }
        status: { type: string, enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, PARTIAL, CANCELLED] }
        progressPercent: { type: integer, minimum: 0, maximum: 100 }
        errorCode: { type: string, nullable: true }
        errorMessage: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Kb:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string, nullable: true }
        visibility: { type: string, enum: [PRIVATE, TENANT] }
        status: { type: string, enum: [ACTIVE, ARCHIVED, DELETING, DELETED] }
        dataRegion: { type: string, description: "创建时继承租户区域（C7 决策）" }
        indexProfileId: { type: integer, format: int64 }
        metadataSchemaId: { type: integer, format: int64, nullable: true }
        retentionPolicyId: { type: integer, format: int64, nullable: true }
        requiresReview: { type: boolean }
        ocrEnabled: { type: boolean }
        activeIndexBuildId: { type: integer, format: int64, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    KbCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 128 }
        description: { type: string, maxLength: 1024 }
        visibility: { type: string, enum: [PRIVATE, TENANT], default: PRIVATE }
        indexProfileId: { type: integer, format: int64, description: "必须指向 ACTIVE 的不可变 index profile" }
        metadataSchemaId: { type: integer, format: int64, nullable: true }
        retentionPolicyId: { type: integer, format: int64, nullable: true }
        requiresReview: { type: boolean, default: true }
        ocrEnabled: { type: boolean, default: true }
        dataRegion: { type: string, description: "缺省继承租户；一经创建不可改" }
    KbUpdateRequest:
      type: object
      properties:
        name: { type: string, maxLength: 128 }
        description: { type: string, maxLength: 1024 }
        visibility: { type: string, enum: [PRIVATE, TENANT] }
        indexProfileId: { type: integer, format: int64, description: "变更将触发新 index build + alias 切换" }
        requiresReview: { type: boolean }
        ocrEnabled: { type: boolean }
    KbMember:
      type: object
      properties:
        userId: { type: integer, format: int64 }
        displayName: { type: string }
        role: { type: string, enum: [OWNER, EDITOR, VIEWER] }
        joinedAt: { type: string, format: date-time }
    KbMemberRequest:
      type: object
      required: [userId, role]
      properties:
        userId: { type: integer, format: int64 }
        role: { type: string, enum: [OWNER, EDITOR, VIEWER] }
    IndexBuild:
      type: object
      properties:
        id: { type: integer, format: int64 }
        profileVersion: { type: integer }
        status: { type: string, enum: [QUEUED, BUILDING, VALIDATING, READY, PUBLISHED, FAILED, RETIRED] }
        inputDocuments: { type: integer }
        chunkCount: { type: integer }
        failedCount: { type: integer }
        qualityGate: { type: object, nullable: true }
        errorCode: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        publishedAt: { type: string, format: date-time, nullable: true }
...
  responses:
...
    NotFound:
      description: 资源不存在或不可见（404）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: 并发/状态机/幂等冲突（409）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/v1/kbs:
    get:
      tags: [knowledge-bases]
      summary: 当前用户可访问的知识库列表（分页）
      parameters:
        - { $ref: "#/components/parameters/page" }
        - { $ref: "#/components/parameters/size" }
      responses:
        "200":
          description: 知识库分页
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/PageData" }
    post:
      tags: [knowledge-bases]
      summary: 创建知识库
      parameters:
        - { $ref: "#/components/parameters/idempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KbCreateRequest" }
      responses:
        "201":
          description: 创建成功（创建者为 OWNER）
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Kb" }
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/kbs/{kbId}:
    parameters:
      - { $ref: "#/components/parameters/kbPath" }
    get:
      tags: [knowledge-bases]
      summary: 知识库详情
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Kb" }
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [knowledge-bases]
      summary: 更新知识库配置
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KbUpdateRequest" }
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Kb" }
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [knowledge-bases]
      summary: 软删除知识库（DELETING → 异步处置）
      parameters:
        - { $ref: "#/components/parameters/idempotencyKey" }
      responses:
        "202":
          description: 已接受删除，返回处置任务
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Task" }
  /api/v1/kbs/{kbId}/clone:
    post:
      tags: [knowledge-bases]
      summary: 异步克隆知识库（配置/ACL 策略复制，文档重建索引）
      parameters:
        - { $ref: "#/components/parameters/idempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 128 }
      responses:
        "202":
          description: 克隆任务已创建
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Task" }
  /api/v1/kbs/{kbId}/members:
    get:
      tags: [knowledge-bases]
      summary: 成员列表
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/KbMember" }
    post:
      tags: [knowledge-bases]
      summary: 添加/更新成员角色
      parameters:
        - { $ref: "#/components/parameters/idempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KbMemberRequest" }
      responses:
        "200":
          description: 成员已加入或角色已更新（触发策略版本增长）
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/KbMember" }
  /api/v1/kbs/{kbId}/members/{userId}:
    parameters:
      - { $ref: "#/components/parameters/kbPath" }
      - name: userId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    delete:
      tags: [knowledge-bases]
      summary: 移除成员（立即收紧权限，索引异步收敛不得放权）
      responses:
        "204":
          description: 已移除
  /api/v1/kbs/{kbId}/index-builds:
    get:
      tags: [knowledge-bases]
      summary: 索引构建历史
      parameters:
        - { $ref: "#/components/parameters/page" }
        - { $ref: "#/components/parameters/size" }
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/IndexBuild" }
    post:
      tags: [knowledge-bases]
      summary: 触发全量重建（新 profile/换模时使用；alias 验证后才原子切换）
      parameters:
        - { $ref: "#/components/parameters/idempotencyKey" }
      responses:
        "202":
          description: 索引构建任务已创建
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Task" }
  /api/v1/index-builds/{buildId}:
    parameters:
      - name: buildId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      tags: [knowledge-bases]
      summary: 索引构建状态
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/IndexBuild" }
...
  /api/v1/tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [governance]
      summary: 通用异步任务状态（进度/错误/取消/重试）
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Task" }
    post:
      tags: [governance]
      summary: 取消任务（仅 RUNNING/QUEUED）
      responses:
        "202":
          description: 已请求取消
...
  /api/v1/kbs/{kbId}/archive:
    post:
      tags: [knowledge-bases]
      summary: 归档知识库（进入 ARCHIVED，停止检索参与）
      parameters:
        - { $ref: "#/components/parameters/kbPath" }
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/ApiResponse" }
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Kb" }
...