# service/src/main/java/com/ragkb/service/modules/knowledge/persistence/entity/IndexBuild.java
package com.ragkb.service.modules.knowledge.persistence.entity;

import com.baomidou.mybatisplus.annotation.IdType;
import com.baomidou.mybatisplus.annotation.TableId;
import com.baomidou.mybatisplus.annotation.TableName;
import java.time.Instant;
import java.io.Serializable;

/**
 * {@code index_build} 表实体 —— MyBatis-Plus 骨架（对齐 {@code deploy/ddl/init.sql}）。
 *
 * <p>约定：主键 {@code BIGINT GENERATED BY DEFAULT AS IDENTITY} → {@link IdType#AUTO}；
 * {@code row_version} 乐观锁 → {@link Version}；snake_case 列自动映射 camelCase 属性
 * （map-underscore-to-camel-case）；时间戳用 {@link Instant}。
 *
 * <p>⚠️ 骨架：JSONB 列当前映射为 {@code String}，业务实现时切换
 * {@code @TableField(typeHandler = JacksonTypeHandler.class)} +
 * {@code @TableName(autoResultMap = true)}；insert 时 null 字段不落 SQL，
 * 由数据库默认值（now() 等）填充。本类不含业务逻辑。
 */
@TableName("index_build")
public class IndexBuild implements Serializable {

    @TableId(type = IdType.AUTO)
    private Long id;

    private Long tenantId;

    private Long kbId;

    private Long indexProfileId;

    private Integer buildNo;

    private String physicalName;

    private String readAlias;

    private String status;

    private String idempotencyKey;

    private Long documentCount;

    private Long chunkCount;

    private Long failedCount;

    private String qualityReport;

    private String errorCode;

    private String errorDetail;

    private Instant queuedAt;

    private Instant startedAt;

    private Instant finishedAt;

    private Instant publishedAt;

    private Instant retiredAt;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Long getTenantId() {
        return tenantId;
    }

    public void setTenantId(Long tenantId) {
        this.tenantId = tenantId;
    }

    public Long getKbId() {
        return kbId;
    }

    public void setKbId(Long kbId) {
        this.kbId = kbId;
    }

    public Long getIndexProfileId() {
        return indexProfileId;
    }

    public void setIndexProfileId(Long indexProfileId) {
        this.indexProfileId = indexProfileId;
    }

    public Integer getBuildNo() {
        return buildNo;
    }

    public void setBuildNo(Integer buildNo) {
        this.buildNo = buildNo;
    }

    public String getPhysicalName() {
        return physicalName;
    }

    public void setPhysicalName(String physicalName) {
        this.physicalName = physicalName;
    }

    public String getReadAlias() {
        return readAlias;
    }

    public void setReadAlias(String readAlias) {
        this.readAlias = readAlias;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }

    public String getIdempotencyKey() {
        return idempotencyKey;
    }

    public void setIdempotencyKey(String idempotencyKey) {
        this.idempotencyKey = idempotencyKey;
    }

    public Long getDocumentCount() {
        return documentCount;
    }

    public void setDocumentCount(Long documentCount) {
        this.documentCount = documentCount;
    }

    public Long getChunkCount() {
        return chunkCount;
    }

    public void setChunkCount(Long chunkCount) {
        this.chunkCount = chunkCount;
    }

    public Long getFailedCount() {
        return failedCount;
    }

    public void setFailedCount(Long failedCount) {
        this.failedCount = failedCount;
    }

    public String getQualityReport() {
        return qualityReport;
    }

    public void setQualityReport(String qualityReport) {
        this.qualityReport = qualityReport;
    }

    public String getErrorCode() {
        return errorCode;
    }

    public void setErrorCode(String errorCode) {
        this.errorCode = errorCode;
    }

    public String getErrorDetail() {
        return errorDetail;
    }

    public void setErrorDetail(String errorDetail) {
        this.errorDetail = errorDetail;
    }

    public Instant getQueuedAt() {
        return queuedAt;
    }

    public void setQueuedAt(Instant queuedAt) {
        this.queuedAt = queuedAt;
    }

    public Instant getStartedAt() {
        return startedAt;
    }

    public void setStartedAt(Instant startedAt) {
        this.startedAt = startedAt;
    }

    public Instant getFinishedAt() {
        return finishedAt;
    }

    public void setFinishedAt(Instant finishedAt) {
        this.finishedAt = finishedAt;
    }

    public Instant getPublishedAt() {
        return publishedAt;
    }

    public void setPublishedAt(Instant publishedAt) {
        this.publishedAt = publishedAt;
    }

    public Instant getRetiredAt() {
        return retiredAt;
    }

    public void setRetiredAt(Instant retiredAt) {
        this.retiredAt = retiredAt;
    }
}